+++ stories/index.tsx
@@ -44,3 +44,38 @@ storiesOf('breakpoint-observer', module).add('Default Breakpoint', () => (
     </BreakpointObserver>
   </React.Fragment>
 ));
+
+storiesOf('breakpoint-observer', module).add('Callback Function', () => {
+  const myCallback = breakpoint =>
+    console.log(`The current breakpoint is ${breakpoint}!`);
+
+  return (
+    <React.Fragment>
+      <h2>Callback function</h2>
+      <p>
+        The <code>{`<BreakpointObserver />`}</code> can also be used with a
+        callback function that receives the current breakpoint. This can then be
+        stored in your application state. This way you do not actually supply
+        any children to <code>{`<BreakpointObserver />`}</code> and can just
+        include it once in your application's top level.
+      </p>
+      <pre>
+        <code>
+          {`const myCallback = breakpoint =>
+    console.log("The current breakpoint is {breakpoint}!");
+
+<BreakpointObserver
+  breakpoints={{ mobile: 0, tablet: 768 }}
+  defaultBreakpoint={'tablet'}
+  callback={myCallback}
+/>`}
+        </code>
+      </pre>
+      <BreakpointObserver
+        breakpoints={{ mobile: 0, tablet: 768 }}
+        defaultBreakpoint={'tablet'}
+        callback={myCallback}
+      />
+    </React.Fragment>
+  );
+});