+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2018-03-23
+### Added
+- Add support for a callback function that receives the current breakpoint
+### Changed
+- children are now optional, so you can keep a single global `<BreakpointObserver />` and sync the breakpoint into your application state with the callback function
+
 ## [0.3.0] - 2018-03-23
 ### Removed
 - Remove `window.innerWidth` support to focus on breakpoints only